refactor: cursor 페이지네이션 조회 로직 수정
