+++ src/main/java/ddingdong/ddingdongBE/domain/feed/service/FacadeClubFeedServiceImpl.java
@@ -76,35 +76,24 @@ public class FacadeClubFeedServiceImpl implements FacadeClubFeedService {
         if (feedPage == null) {
             return MyFeedPageQuery.createEmpty();
         }
-        List<Feed> completeFeeds = feedPage.getContent().stream().filter(this::isComplete).toList();
-
-        List<FeedListQuery> feedListQueries = completeFeeds.stream().map(feedFileService::extractFeedThumbnailInfo).toList();
+        List<Feed> completeFeeds = feedPage.getContent();
+        List<FeedListQuery> feedListQueries = completeFeeds.stream()
+                .map(feedFileService::extractFeedThumbnailInfo)
+                .toList();
         PagingQuery pagingQuery = PagingQuery.of(currentCursorId, completeFeeds, feedPage.hasNext());
 
         return MyFeedPageQuery.of(feedListQueries, pagingQuery);
-    }
-
-    private boolean isComplete(Feed feed) {
-        if (feed.isImage()) {
-            return true;
-        }
-
-        VodProcessingJob vodProcessingJob = vodProcessingJobService.findByVideoFeedId(feed.getId());
-        if (vodProcessingJob == null) {
-            return false;
-        }
-        return vodProcessingJob.isCompleteNotification();
     }
 
     private void checkVodProcessingJobAndNotify(Feed feed) {
         VodProcessingJob vodProcessingJob = vodProcessingJobService.findByVideoFeedId(feed.getId());
         if (vodProcessingJob != null && vodProcessingJob.isPossibleNotify()) {
             SseEvent<SseVodProcessingNotificationDto> sseEvent = SseEvent.of(
-                "vod-processing",
-                new SseVodProcessingNotificationDto(
-                    vodProcessingJob.getVodProcessingNotification().getId(),
-                    vodProcessingJob.getConvertJobStatus()),
-                LocalDateTime.now()
+                    "vod-processing",
+                    new SseVodProcessingNotificationDto(
+                            vodProcessingJob.getVodProcessingNotification().getId(),
+                            vodProcessingJob.getConvertJobStatus()),
+                    LocalDateTime.now()
             );
             sseConnectionService.sendVodProcessingNotification(vodProcessingJob, sseEvent);
         }
